SC-1906: review changes
